requirements.txt
Django==1.11.3
django-crispy-forms==1.7.0
django-formtools==2.1
django-haystack==2.6.1
django-ranged-response==0.2.0
django-simple-captcha==0.5.6
future==0.16.0
httplib2==0.9.2
jieba==0.39
Markdown==2.6.9
olefile==0.44
Pillow==4.3.0
PyMySQL==0.7.11
pytz==2017.3
six==1.11.0
Whoosh==2.7.4
